+++ src/OFDate.m
@@ -538,7 +538,7 @@ tmAndTzToTime(struct tm *tm, int16_t *tz)
 	return true;
 }
 
-- (uint32_t)hash
+- (unsigned long)hash
 {
 	uint32_t hash;
 	double tmp;
